tdf#56258 svx SetOutline: convert to toggle
The "Outline Format" toolbar button is normally hidden.
It didn't check state like the bullet and toggle numbering do.
This patch adds that - so it lights up when the text
contains a numbering list that matches a valid outline choice.

I also added a toggle aspect to the button.

Toggling ON is not great. It works fine if there is no
numbering yet, but if you are in an outline and toggle off,
it won't toggle back on to the same list - it starts a new one.
I don't think there is much that can be done about that,
since the removal delete the level information too,
and outlines are all about multiple levels of numbering.
The proper response to an errant
toggle off is (multiple) undo(s), not toggle back on.
(In the ideal world, it would popup the dropdown choices).

Toggling ON does nothing when there are already bullets
or numbering. The proper response is to use the dropdown
if the desire is to change to an outline format.

Toggling OFF is only possible when we are on a valid outline.
This is extremely important because otherwise it isn't clear
how to remove Outline Numbering's mix of bullets and numbers.
(When the selection contains one list
with both bullet and number levels,
the number and bullet buttons do nothing.)

Although not perfect, this is not a bad user experience,
so I'm going to push the full patch to gerrit.

There is one known obscure case that troubles all methods.
If the selection spans multiple list definitions and ends
on an un-numbered paragraph, then the toggle-on will happen.

It is possible to just not do any toggling
and ALWAYS show the dropdown action,
but that isn't very consistent with the other related buttons.

Change-Id: I6d927604e49711f1da8549be9f3dbf8b8133a06b
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/145061
Tested-by: Jenkins
Reviewed-by: Justin Luth <jluth@mail.com>