tdf#115965 sw RemoveBullet: do not light-up button when NumRule
Sure, most of the time it was "off" when there was either a
bullet or a number list. However, when the selection had
both a bullet and a number list/level, then it would light up.
So just make it simple. If there is a numrule in the selection,
then it shouldn't be "on".

There is still one obscure case it lights up funny.
If more than one list is selected, and the selection ends
on a non-numbered paragraph, then RemoveBullet
will still light up - as if there is no numbering in the selection.

Change-Id: I5c37d53206918068b2994fde9774541ea988a623
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/145095
Tested-by: Jenkins
Reviewed-by: Justin Luth <jluth@mail.com>