sw lok: get all refmarks: sort the refmarks array
This was using the order used by GetItemSurrogates(), which is mostly by
pointer address.

We know that these pool items have a text attribute, which have a node
and a content index, sort by that.

In theory two refmark may start at the same doc model position, but
that's never the case when using Zotero, so don't worry about that for
now.

For a document with 5 refmarks, the original order I got was 2-3-5-4-1,
now it's property 1-2-3-4-5.

Change-Id: I2768845414bd36afca91ec02a0f3364c246ddfd9
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/145145
Reviewed-by: Miklos Vajna <vmiklos@collabora.com>
Tested-by: Jenkins