tdf#56258 i18npool en-US: extend outline 1.a) to 4 levels
Change 1. a) -> 1) a) 1) A)

This is intended as a clean, modern style of outline
that has good visual spacing in LibreOffice.

This differentiates it a bit more from the first option,
and extends the numbered level to the typical depth of 4,
while still keeping a logical suffix pattern
and maintaining the general "theme" of the previous definition.

Yes, it is a bit odd to have a capital A following
a smaller case a. However, this is at the fourth
level, which isn't reached very often,
and I wanted to keep a choice that focuses on a small a.

Surprisingly, with the default 12pt/Liberation font, a
10) still fits in the allotted space before the tabstop,
so in most cases this outline will have proper visual spacing
all the way up to 99).

There are plenty of choices that start with 1 followed by a dot,
so I used this opportunity to create the only list
that starts with a ")" suffix. Google docs for example has a
1) a) i) choice, but for modern lists I don't think
we want to use roman numerals.

Change-Id: I34725d6aa113831eaeaa88c295b3b3f1294f8998
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/145155
Tested-by: Jenkins
Reviewed-by: Justin Luth <jluth@mail.com>